+++ lib/karafka/web/pro/ui/controllers/consumers/consumers_controller.rb
@@ -71,9 +71,7 @@ module Karafka
                 current_state = Models::ConsumersState.current!
                 @process = Models::Process.find(current_state, process_id)
 
-                return render if @process.schema_compatible?
-
-                raise Errors::Ui::IncompatibleSchemaError
+                render
               end
 
               # @param process_id [String] id of the process we're interested in