+++ spec/lib/karafka/web/errors_spec.rb
@@ -31,12 +31,6 @@ RSpec.describe_current do
 
       specify { expect(error).to be < described_class::BaseError }
     end
-
-    describe 'IncompatibleSchemaError' do
-      subject(:error) { described_class::Processing::IncompatibleSchemaError }
-
-      specify { expect(error).to be < described_class::BaseError }
-    end
   end
 
   context 'when in Ui namespace' do
@@ -54,12 +48,6 @@ RSpec.describe_current do
 
     describe 'ForbiddenError' do
       subject(:error) { described_class::Ui::ForbiddenError }
-
-      specify { expect(error).to be < described_class::BaseError }
-    end
-
-    describe 'IncompatibleSchemaError' do
-      subject(:error) { described_class::Ui::IncompatibleSchemaError }
 
       specify { expect(error).to be < described_class::BaseError }
     end